typo fixed.

+++ financial-helper-cli/src/main/java/dev/kerim/fettahoglu/financial_helper_cli/repo/StockRepository.java
@@ -3,6 +3,6 @@ package dev.kerim.fettahoglu.financial_helper_cli.repo;
 import dev.kerim.fettahoglu.financial_helper_cli.entity.Stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockRepository extends JpaRepository<Stock, Integer> {
+public interface StockRepository extends JpaRepository<Stock, Long> {
 
 }
